+++ examples/resources/event_gateway.tf
@@ -1,0 +1,3 @@
+resource "konnect_event_gateway" "demo" {
+  name = "tf-event-gateway"
+}