+++ internal/provider/gatewaypluginkafkaupstream_resource_sdk.go
@@ -81,7 +81,7 @@ func (r *GatewayPluginKafkaUpstreamResourceModel) RefreshFromSharedKafkaUpstream
 		if resp.Config.SchemaRegistry == nil {
 			r.Config.SchemaRegistry = nil
 		} else {
-			r.Config.SchemaRegistry = &tfTypes.SchemaRegistry{}
+			r.Config.SchemaRegistry = &tfTypes.ConfluentPluginSchemaRegistry{}
 			if resp.Config.SchemaRegistry.Confluent == nil {
 				r.Config.SchemaRegistry.Confluent = nil
 			} else {
@@ -93,7 +93,7 @@ func (r *GatewayPluginKafkaUpstreamResourceModel) RefreshFromSharedKafkaUpstream
 					if resp.Config.SchemaRegistry.Confluent.Authentication.Basic == nil {
 						r.Config.SchemaRegistry.Confluent.Authentication.Basic = nil
 					} else {
-						r.Config.SchemaRegistry.Confluent.Authentication.Basic = &tfTypes.Basic{}
+						r.Config.SchemaRegistry.Confluent.Authentication.Basic = &tfTypes.BackendClusterAuthenticationSaslPlain{}
 						r.Config.SchemaRegistry.Confluent.Authentication.Basic.Password = types.StringValue(resp.Config.SchemaRegistry.Confluent.Authentication.Basic.Password)
 						r.Config.SchemaRegistry.Confluent.Authentication.Basic.Username = types.StringValue(resp.Config.SchemaRegistry.Confluent.Authentication.Basic.Username)
 					}