+++ internal/provider/types/app_auth_strategy_config_key_auth.go
@@ -8,4 +8,5 @@ import (
 
 type AppAuthStrategyConfigKeyAuth struct {
 	KeyNames []types.String `tfsdk:"key_names"`
+	TTL      *TTL           `tfsdk:"ttl"`
 }